Merged PR 2573: Ping Protect SDK integration - 20627
The Ping Protect SDK should be successfully integrated and initialized in the application.
The getData() method must be invoked during user login, and collected the data as deviceInformation.
pass request parameters such as ipAddress, userAgent, and deviceInformation to the loginAuthentication API. Ensure that all related methods are mocked, and the parameters are correctly passed to all test cases to ensure they pass successfully.

As part of Ping Protect Implementation:

scenario 1:  If a user is found to have too high of a risk score after entering their username and password (Access Denied) and the returned response error code is PP-600 it will redirect the user to the public error page: https://www.bcbst.com/error/
scenario 2: if the user failed to determine the risk level after enter username and password and the returned response error code is PP-601 it will redirect the user to the public error page: https://www.bcbst.com/error/

Related work items: #20627