Merged PR 3233: US-18769 Date Picker Calendar Component
As part of the Development for Calender Datepicker created reusable component for calender that can be integrated where ever it is required.
**covered the scenarios are**
User can able to select the date from calender
User can able to enter date manually and while entering month and date slashes should be automatically added accordingly
calender is displayed in the form of tool tip
Followed all the font sizes and font familes and changed colors of dates based on the invison to match the look and feel
in the current month calender next month dates are disabled
- Allow selection from a date range if required.
- Validation for incorrect dates.

as of now for error scenario if user enters the date before the year 1900 showing error message but actual business logic has not been implemented. Ly mentioned that they will create separate story for that.

Related work items: #18769