Merged PR 5698: US-15752 Share My Information - API Integration
Share Plan Information with different levels of access on my plan.

plan information includes
Name
Date of birth
Level of access should be Basic Access or Full Access or No Access
If a member is a minor dependent (13-17 years old) then the logged in member shouldn't be able to share access with them. There will be a box that says: This is a minor dependent. Sharing permissions aren’t applicable with this account.

Related work items: #15752