Merged PR 7680: Update Portal UI to Remove Unnecessary Component
As developer, I want to update the member portal UI with the latest requirement.

If a member arrives on the dashboard and scrolls down to the Benefits & Coverage component, they should not see a search field. (This was already completed by Brady as part of 66994)

If a member arrives on the Profile Settings page, we should NOT see the Update button under phone number.

If a member arrives on the Sharing and Permission pages, we should NOT see the Third Party Sharing option.