Merged PR 9513: M3P - Terms & Conditions - backend integration
BUG - 77194 : added impl for the missed requirement on backend integration for M3P terms and conditions before redirecting to sso pages.

BUG - 77439 : Added visibility rule check on benefit pages to show the spending account side comp, fixed the url on the balances page for the redirection & applied the correct CSS for this comp.

Related work items: #77194, #77439